Add express-jwt middleware function
In this commit, we added the express-jwt middleware function to the authJwt function. The authJwt function returns a middleware function that verifies the JWT in incoming requests using the secret and algorithms options. The expressJwt function is imported from the express-jwt module using the require() function. The middleware function can be applied to routes using the app.use() method.