Add JWT authentication middleware
In this commit, we added the authJwt middleware function to the Express application. The authJwt function is defined in the ./helper/jwt.js file and is used to authenticate requests using a JSON Web Token (JWT). The authJwt function returns a middleware function that verifies the JWT in incoming requests using the secret and algorithms options.

The authJwt middleware function is imported into the main application file using the require() function:

The middleware function can be applied to specific routes or to all routes using the app.use() method: