Commit message: Added error handling middleware to the app
Changes Made:

Added the 'errorHandler' middleware to the application using the 'app.use()' method.
This middleware will now handle any errors that occur during the request/response cycle in the application.
The 'app.use()' method registers middleware functions that can be used for all HTTP methods and all routes in the application.