Commit message: Set the API URL as an environment variable
Changes Made:

Created a constant named 'api' that stores the value of the environment variable 'API_URL'.
The 'API_URL' environment variable can be used to store the base URL of the API server that the application will communicate with.
Storing the API URL in an environment variable allows for easy configuration and deployment of the application to different environments.