Commit message: Added isRevoked property to authentication middleware
Changes Made:

Added a new property called 'isRevoked' to the authentication middleware.
The 'isRevoked' property is a function that can be used to revoke the authentication token for a user.
This function can be used to implement functionality such as logout or revoking tokens for specific users.
The 'isRevoked' property will be set to the 'isRevoked' function that is defined elsewhere in the application.