Commit message: Added isRevoked function to revoke tokens for non-admin users
Changes Made:

Added a new asynchronous function called 'isRevoked' that takes in the request, the token payload, and a callback function called 'done'.
The function checks whether the user specified in the token payload is an administrator or not.
If the user is not an administrator, the function calls the 'done' callback with the second parameter set to 'true' to indicate that the token should be revoked.
If the user is an administrator, the function calls the 'done' callback without any parameters to indicate that the token is valid.
This function can be used as the 'isRevoked' property in the authentication middleware to revoke tokens for non-administrator users.