This code defines a GET route using the Express router object.
The purpose of this route is to retrieve a list of orders from the database.

router.get('/', async (req, res) => {...}): This defines a GET route on the root URL path ('/').

const orderList = await Order.find().populate('user', 'name').sort({'dateOrdered': -1});: This line of code queries the database using the Order model to retrieve all orders, with the user field populated with only the name property. The results are sorted by dateOrdered in descending order. The await keyword is used to indicate that this is an asynchronous operation and the code will wait for the result before continuing.

if(!orderList) {...}: This block of code checks whether orderList is falsy, which would indicate that there was an error retrieving the orders from the database. If this is the case, it sends an error response with status code 500 and a JSON object with a success property set to false.

res.send(orderList);: This sends a successful response with the orderList object retrieved from the database as the response body.