feat: set up multer storage for file uploads
This commit adds a multer instance and sets up its storage object with a destination and filename property. The destination property specifies the directory where the uploaded file will be stored, while the filename property specifies the name of the uploaded file. Uploaded files must have a valid MIME type; if a file's MIME type is invalid, an error will be returned to the client. The multer instance can be used with an HTTP POST request to upload files to the server.